App URLs to default settings and random mac
Moved the app URLs to default settings so they can be easily configured for white label.
Separated the Windows App Link and the QR Code so they now use different passwords, so you can click the link and scan the code to activate 2 separate devices.
Changed MAC generation from just being based on the extension to random, starting with 22-22 because that's in a "local admin" range.